split message
